BasisWriter: redesign using Strategy pattern
This commit redesigns BasisWriter in a backward-compatible way using a
Strategy pattern so that it can be used in callbacks within
HyperreducedModelGenerator.

The main modification is to add an enumerated type and a field of that
enumerated type to signify which private method
BasisWriter::writeBasis() should delegate to when called. A private
method for each of the values in the enumerated type is added so that
writeBasis() does the "right thing" when called.